Basic Building Preview
The basic form of the preview is now part of the C++ side of project.
All the states are structured as gameplay Tags, all the tiles are stored unanymously in a one single database.
Thanks to some restructurizations and adjustments compared to the previous, Blueprint version of the logic, it now takes less lines of bloated code to achieve the tasks that in the future will most likely become a routine.
Yet to add rotation control over the Preview Mesh, different Preview States and the Snapping System algorythm.